pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibes"
version = "0.1.0"
description = "Twitter-like SPA for interacting with coding agents via ACP protocol"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-aiohttp>=1.0.0",
]

[project.scripts]
vibes = "vibes.app:main"

[project.urls]
Homepage = "https://github.com/rcarmo/vibes"
Repository = "https://github.com/rcarmo/vibes"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
vibes = ["static/**/*"]
